Query editor PoC
