+++ docs/source/api/msticpy.config.query_editor.rst
@@ -1,0 +1,7 @@
+msticpy.config.query\_editor module
+===================================
+
+.. automodule:: msticpy.config.query_editor
+   :members:
+   :undoc-members:
+   :show-inheritance:
