+++ docs/source/api/msticpy.data.core.query_template.rst
@@ -1,0 +1,7 @@
+msticpy.data.core.query\_template module
+========================================
+
+.. automodule:: msticpy.data.core.query_template
+   :members:
+   :undoc-members:
+   :show-inheritance:
